Day 11 - Fraud Detection Engine Added

+++ src/main/java/com/bhavy/banking_fraud_system/entity/Transaction.java
@@ -26,7 +26,7 @@ public class Transaction {
 
     private LocalDateTime transactionTime;
 
-    private Double fraudScore;
+    private Integer fraudScore;
 
     private Boolean isFraud;
